refactor: Move composables out of MainActivity.kt
Files ending in `Activity` are not configured to have unit test coverage.
