+++ test-wrapper/src/test/snapshots/images/uk.gov.onelogin.criorchestrator.testwrapper_ScreenshotTest_screenshot[2.uk.gov.onelogin.criorchestrator.testwrapper.MainContentKt_MainContentPreview]_light.png
